#5: Rename `Benchmark` to `session`, `TestCase` to `Prompt`
