--- src/promptum/benchmark/__init__.py
@@ -1,8 +1,0 @@
-from promptum.benchmark.benchmark import Benchmark
-from promptum.benchmark.report import Report
-from promptum.benchmark.result import TestResult
-from promptum.benchmark.runner import Runner
-from promptum.benchmark.summary import Summary
-from promptum.benchmark.test_case import TestCase
-
-__all__ = ["Benchmark", "Report", "Runner", "Summary", "TestCase", "TestResult"]
