+++ src/promptum/session/summary.py
